Test for stbt-run uses stbt from development, not installed, location.
By using "stbt-run" instead of "stbt run", the test picks up the program
from the source directory (PATH is set up in the recipe for "make
check"). It used to pick up "stbt" from /usr/local/bin.

To avoid relying on the default stbt.conf, which may or may not be
installed, we also pass in defaults to all options, even those we don't
use ("--control", in this case).

In future we may want to install these tests as "stbt selfcheck", in
which case we'll want them to use the installed "stbt" ("stbt-run" etc
won't be in the path; they're installed to $prefix/lib/stbt/). But we'll
worry about that bridge when we reach it.