Fix receiving keypresses from VirtualRemote
7537b5de fixed the read_records doctest but broke the actual use of
read_records in production.  I had expected that

    socket.makefile(bufsize=0).read(x)

be equivalent to

    socket.recv(x)

but it turns out not to be the case.  This seems to be fixed in later
versions of Python[1].  To work around this we convert the StringIO used
in the tests to be socket-like rather than converting the socket used in
production to be file-like.

[1] http://stackoverflow.com/questions/6946629/can-i-get-a-socket-makefile-to-have-the-same-read-semantics-as-a-regular-file