Make the video sink be configurable with --sink-pipeline
This allows running stbt and its tests on OS X and on Linux VMs (on OS X
there is no xvimagesink, only ximagesink without the v; on my Linux VM
the xvimagesink brings up a blank window with no video, whereas
ximagesink does work).

The default value remains the same (xvimagesink sync=false). We might
choose to change it to:

* autovideosink -- this introduces a dependency on gst-plugins-good, and
  doesn't work on OS X (where it just hangs with no error message nor
  window; presumably it's trying to find xvimagesink), and it doesn't
  work on my Linux VM (where it brings up the xvimagesink window, which
  doesn't have any video, just a blank window). Also we'd need to test
  that autovideosink passes on properties like "sync=false" to the
  underlying sink.

* ximagesink -- less performant than xvimagesink, but works for everyone
  and thus is arguably a better out-of-the-box experience for the first-
  time downloader who just wants 'make check' to show some cool demos.