Cleanup test runner
This brings a clear distinction between the part of run-tests.sh that
is defining helper functions for the tests, and the part that is running
each test.

By moving test_wait_for_match into its own function I get the
flexibility of doing anything I want in that function, without being
constrained by having to do everything in the parameters to run().