Add "is-live=true" to default videotestsrc
Without "is-live=true", the timing information from videotestsrc is
extremely accelerated, so in the test you don't get to see what's
happening on screen. With consecutive_matches=24, this test should
now stay on screen for an entire second.

tests/stbt.conf is added because it will always over-ride any site- and
user-local configuration (the tests are run from tests/, and a stbt.conf
in the current directory over-rides all other config files).