Documentation: Clarify software requirements
gst-plugins-base is needed for pretty much everything we use:
videotestsrc, ffmpegcolorspace, {x,xv}imagesink...

gst-plugins-good is needed for pngenc (which we use to take screenshots
of failed tests), and v4l2src (the Video-for-Linux-2 source for the
Hauppauge capture device).

pygst is the python gstreamer bindings.

pygtk we use only for its main loop, which delivers us the gstreamer
signals/messages.