wait_for_match fails immediately if it can't load the template image
The templatematch element doesn't raise an error but a warning, because
(for some applications) you might not want to stop the whole pipeline
just because of templatematch.

This fix requires the following patch to
gst_template_match_load_template in
gst-plugins-bad/ext/opencv/gsttemplatematch.c:
https://bugzilla.gnome.org/show_bug.cgi?id=678485#a6

I don't know how better to test that the warning message is, in fact,
due to failure to load the template image. The approach in this commit
will fail if the warning message is changed or translated. Perhaps we
can check the RESOURCE and FAILED codes, plus the image filename
anywhere in the debug field.