Selftests: Move python stbt scripts into test-stbt-run.sh
There was way too much indirection to add a single test (or understand
an existing one):

1. Invoke the shell function in run-tests.sh;
2. define it in (say) test-stbt-run.sh, where it merely passes the stbt
   script to stbt-run;
3. the stbt script itself in (say) test-wait_for_match.py.

Now you still have to mention the test twice: Define it in
test-stbt-run.sh, and run it from test-stbt-run.sh. I want to keep it
this way, rather than saying "run all shell functions starting with
test_", because I want to control the order the tests are run: First
test your gstreamer installation (and stop as soon as the first test
fails so we don't overwhelm the first-time downloader who ran "make
check").

But at least now you can understand the entire test case just by looking
at a single place: The test_wait_for_match function in test-stbt-run.sh.

If the test fails, the stbt script will be left in $scratchdir/test.py
(the test runner will tell you the location of $scratchdir) so you can
run it manually with "stbt-run $scratchdir/test.py" to investigate.

In case you're wondering, the "<<-EOF" to strip leading tabs in the
here-doc is in posix:
http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_07_04