stbt-record: Avoid extra keypress after Control-C
When you finish recording a stbt script, the standard way of ending the
recording is to press Control-C in the terminal window. After the
Control-C, stbt-record would print the last-pressed key again (it didn't
actually make it into the recorded script, but it did show on the debug
output, causing confusion).