Move Display from stbt-run to stbt.py
This is in preparation for a videotestsrc remote control that will be
used by the selftests: The remote control will need to access Display's
internal details (its source pipeline), so they should be in the same
module.

The order in stbt.py looks like this: gstreamer-related stuff, remote-
control-related stuff, exception hierarchy, miscellaneous.