videotestsrc remote control for the selftests
Use it with `--control=test`. Only works with
`--source-pipeline=videotestsrc`.

A stbt script like `press("18")` will cause the videotestsrc element to
change to pattern 18 (see `gst-inspect videotestsrc` for the available
patterns).

By the way, I changed the uri for the null control to be case
insensitive (so it can be `--control=none` instead of `--control=None`)
for consistency with `--control=test`.