stbt-record: Handle EOF in control-recorder input
End-Of-File is handled just as well as a Control-C is.

This is essential for file-based input: `--control-recorder=file://...`
(although file-based input still won't work well, because we need to
somehow introduce a delay before the first keypress so the gstreamer
pipeline has a chance to start up, and between keypresses so the video
source has a chance to react).

In `file_remote_recorder`, note that file.readline() returns an empty
string *only* when EOF is encountered immediately.