Support LIRC infrared emitters & receivers
LIRC is "Linux Infrared Remote Control", a package for sending &
receiving infrared signals via USB or serial infrared devices (the
emitters are commonly known as "IR blasters").

See the changes to README.rst for configuration instructions.

The format that lircd expects from a client wanting to *send* a signal
is the same as taken by the irsend(1) command-line tool.

The format of messages indicating a *received* signal is:

    <code> <repeat count> <button name> <remote control name>

where <code> is to be ignored. Holding down a key on the remote control
results in multiple lircd messages, with an increasing <repeat count>. I
had to try hard to release the keys quickly enough to only get a single
message, so here I'm ignoring all but the first message (with repeat
count of 0). This will need to be revisited if we need to emulate
holding down a button.

Both formats are documented at
http://www.lirc.org/html/technical.html#applications