stbt-record: Fix local video display
stbt-record would correctly capture screenshots, but wouldn't show the
video on your (the test rig's) screen. This isn't a problem when you
have a monitor connected to the set-top box right in front of you; but
in a large setup the set-top boxes might not be connected to a monitor
at all.

The problem was that we were lacking a colorspace conversion in between
the video source and the X video sink. The selftests (with their
videotestsrc) worked correctly; using xvimagesink also worked correctly,
because it takes the same video/x-raw-yuv produced by our V4L2 source;
but ximagesink only takes video/x-raw-rgb, so it requires the conversion
introduced in this commit.