README/manpage: Fix documented load order of config files
The README and load_defaults (in stbt.py) didn't agree on the order of
loading $STBT_CONFIG_FILE and ./stbt.conf. I chose to keep the current
behaviour and fix the documentation to match, because the selftests rely
on tests/stbt.conf (in the directory from which the selftests are run)
to override any local user configuration.