Lirc & VirtualRemote: Clearer error message on failure to connect
There are several file-like objects involved in stb-tester (the control
socket, the control recorder socket, the script file) so the bare
exception backtraces were proving to be too obscure for immediately
spotting the error.

Note on the socket.error formatting: The received socket.error can
either have errno & strerror set, or message set. Instead of checking
each one, I use str(e) to retrieve the current message.

Then I change the message by changing args *and* strerror, because it
depends on the exact subclass of socket.error as to which of these
attributes is used when printing the exception (a timeout error seems to
use args, whereas a "name or service not known" socket.gaierror uses
strerror).

Note that I modify the exception object in-place and re-raise it, to
preserve the original backtrace.