Ignore templatematch messages received before appsink has received any frames
We have an appsink (self.screenshot in Display.bus_message) that we use
to determine the current timestamp of the stream. We calculate the
wait_for_match timeout from that timestamp.

The very first templatematch messages can arrive before the appsink has
received any frames, which caused the following error:

      File "/usr/local/lib/stbt/stbt.py", line 229, in bus_message
        self.start_timestamp = buf.timestamp
    AttributeError: 'NoneType' object has no attribute 'timestamp'

This error happens during a callback (to Display.bus_message), and
presumably the exception is caught by the callback dispatcher mechanism
(the gtk.main loop), because it didn't stop `stbt run` from proceeding.
This commit merely removes the distracting error message from the logs.