Continue re-starting the source bin if it wasn't available
For example if the video signal didn't return before the 5 second sleep.

Note that the Hauppauge takes a few seconds to initialise itself once it
detects an input signal; starting the pipeline before the Hauppauge is
ready gives the error "libv4l2: error getting pixformat: Bad address".

We need to do this because the above libv4l2 error isn't reported to the
application (see the call to v4l2_fd_open in gst_v4l2_open in
gst-plugins-good/sys/v4l2/v4l2_calls.c).

This doesn't seem to work when the video signal has been unavailable for
too long (more than 1 or 2 attempts at `restart_source_bin`): After
video signal is re-established, when the source bin in restarted the
Hauppauge light turns on but the window stays blue and no frames are
received by the pipeline.