Wrap the source pipeline in a "source bin"
This is in support of future changes to recover from loss of input video
signal.

A gstreamer "bin" is a pipeline (of several gstreamer elements) that can
be treated like a single element.

Note that gst.parse_bin_from_description(_, True) tells it to create the
"ghost pads" -- the pads of the bin as a whole, connected to a pad of a
particular element in the bin. This works when creating the sink bin
because its elements only contain a single unconnected sink pad (the
sink of the left-most queue).

The pipeline of the source bin is outside our control (because it is
configurable) and it might have multiple source pads, so we create the
ghost pad manually.

The source pipeline might not know all of its pads yet: For example
decodebin (used in our source pipeline for the Hauppauge HDPVR capture
device) doesn't know its source pad until it knows the format of the
stream it will be decoding, i.e. until the pipeline has started playing.
When we were using gst.parse_launch, it (presumably) was registering a
handler for decodebin's "new-decoded-pad" signal, and linking the pads
in that handler. Instead of doing that ourselves, we add a capsfilter
element to force decodebin to create a source pad with the specified
caps.