virtual_remote_listen log message includes address as well as port
A common error is to specify "--control-recorder=vr:localhost" when
running the VirtualRemote application from a different host.

(In this case you should use "--control-recorder=vr:" to listen on all
interfaces.)

We already see the value of "--control-recorder" in the log when stbt
starts running; but maybe seeing "localhost" again at the "Waiting for
connection" message will make it more obvious that it is only listening
on localhost.