templatematch: Add second pass to the matching algorithm for robustness
We've had problems where setting the match threshold at 99% failed to
match some frames it should have (probably due to noise introduced by
our video capture device's H.264 encoding), but reducing the threshold
would then give false positives with other frames.

This commit introduces a 2nd pass to the matching algorithm. We start
by substantially lowering the threshold of the 1st pass to 80%; this
should eliminate the false negatives, and introduce many false
positives. Then we run the 2nd pass only on the portion of the frame
that matched in the 1st pass (this minimises the performance effect
of the 2nd pass). The 2nd pass does another templatematch but on the
*edges* of the frame & template images (as produced by the "canny"
edge detection algorithm).

We have some automated tests (that we haven't committed to this
repository yet for reasons outside our control) that confirm this
algorithm better suits our needs.