templatematch: Detect small but significant differences with large template image
With large templates, when only a small part of the source frame differs
from the template image the "match certainty" returned by our algorithm
was high.

The problem is that there is an important difference between a small
difference concentrated in one portion of the image, and the same amount
of difference overall but spread across the image (noise). Increasing
the match certainty threshold to detect small concentrated differences
also rules out images that should match but have a lot of noise.

The new algorithm splits the images into 100x100 pixels and applies the
templatematch certainty threshold to each 100x100 portion; all portions
have to pass the threshold for the entire image to be considered a
match.