gstreamer plugin: Boilerplate C code for stbt-motiondetect element
It's called STBT-motiondetect because we plan to move the templatematch
element from gst-opencv/gst-plugins-bad here to stb-tester (so that we
can make custom modifications to the image matching algorithm that
probably won't be of generic interest); we will call that element
stbt-templatematch so that it won't clash for users with gst-opencv
installed, so the name "stbt-motiondetect" is consistent with that
future element.

The stbt-motiondetect element doesn't actually do anything yet, but all
the GObject boilerplate and the Makefile to build it is in place.

This does compile and work in a gstreamer pipeline:

    GST_PLUGIN_PATH=/path/to/stb-tester/gst \
    gst-launch \
        videotestsrc ! ffmpegcolorspace ! stbt-motiondetect ! \
        ffmpegcolorspace ! ximagesink