Makefile: Cleanup internal order of various targets
I'm aiming for the best order for reading top-to-bottom for someone new
to this project/Makefile.

First the customisation points that can be set on the command line, e.g.

    make install prefix=/usr

Next the project's dependencies, makefile variables (VERSION), and
configuration of the behaviour of `make` (.DELETE_ON_ERROR). I thought
about moving OBJS down to where the details of building the gstreamer
plugin are, but the dist target needs OBJS to have been set.

Next the main targets of interest: all, install, doc, clean, check, dist
(the aim is that the reader can see the main targets available, without
being overwhelmed with details).

Finally the details of building the gstreamer plugin.