jenkins-stbt-run: Shell script to be run by Jenkins
Runs the stbt scripts specified on the command line; creates logfiles
(stdout, stderr, result code, stbt's failure screenshot) as "build
artifacts" in the location expected by Jenkins; and generates an xUnit
xml file with the test results.

jenkins-stbt-run doesn't return a failure exit status if the test
scripts fail; a failure exit status means an error in the script itself.
Instead configure the Jenkins xUnit plugin to consider the job as
"failed" or "unstable" above 0 failing tests.

I didn't find any conclusive documentation on the xUnit xml format; the
closest was this schema:
https://svn.jenkins-ci.org/trunk/hudson/dtkit/dtkit-format/dtkit-junit-model/src/main/resources/com/thalesgroup/dtkit/junit/model/xsd/junit-4.xsd

Jenkins groups the test results by "package" and "classname", which it
takes from the dot-separated names in the "name" attribute of the
"testsuite" xml element.

Note that I don't need to put quotes around parameter (variable)
expansions, even when the variables could contain spaces, because
jenkins-stbt-run modifies IFS so that word-splitting after parameter-
expansion doesn't split on spaces. This is necessary in case the test
script paths themselves contain spaces. I do need the quoting in the
argument to `trap`, because it is expanded once as a string and then
evaluated later.

jenkins-stbt-run cleans out $WORKSPACE before each invocation. In spite
of having configured the Jenkins job to "Wipe out workspace before
build" (under "Source code management" > "Git" > "Advanced") the whole
workspace isn't cleared out; only the git repository containing the test
scripts. So if a test script succeeded, but had failed in a previous
run, the failure screenshot was still lying around and captured by
Jenkins as a "build artifact".

JENKINS CONFIGURATION
=====================

Nodes
-----

Install the Jenkins SSH Slaves plugin.

A "slave" or "node" is a PC with video-capture and remote-control
connections to the set-top box under test. Jenkins will connect to the
slave and run `jenkins-stbt-run` on the slave. Jenkins requires java and
"jenkins.jar" to be running on the slave; the "SSH Slaves" plugin merely
connects to the slave via ssh and installs and runs jenkins.jar, so that
you don't have to do that installation manually.

At the Jenkins index page:

Manage Jenkins > Manage nodes > New Node

* # of executors: 1
* Remote FS root: /home/stb-tester
* Launch method: Launch slave agents on Unix machines via SSH
* Host: the slave's hostname or ip address
* Username: stb-tester
* Private key file: Path (on the Jenkins master) to the ssh key
  (add the corresponding public key to ~stb-tester/.ssh/authorized_keys
  on the slave)

Jobs
----

Create a new Jenkins job (aka project):

* Restrict where this project can be run
  Label expression: the name of the node(s) to run this job

  Note that the entire job (all its stbt scripts) will be run on a single
  slave; this system (currently, at least) doesn't split the job across
  multiple slaves. Specifying multiple slaves (nodes) here means that
  Jenkins will pick one of the available nodes to run this job.

* Build > Execute shell
  Command: /home/stb-tester/jenkins-stbt-run tests-repo/path/to/test_directory

  where "tests-repo" is a repository containing your stbt scripts, which
  will be checked out by Jenkins according to your configuration in the
  job's "Source Code Management" section. "path/to/test_directory" is
  any path within that repo. Don't include "test.py", just the directory
  containing "test.py". You can also specify a parent directory
  containing multiple stbt scripts: jenkins-stbt-run looks for
  **/test.py under the path you specify. You can specify multiple paths
  on this command line.

xUnit
-----

Install the xUnit plugin.

In the job/project:

* Configure > Post-build actions > Publish xUnit test result report; then

* Add > JUnit > JUnit Pattern: results.xml
  (This is the path to the xml file generated by jenkins-stbt-run,
  relative to the Jenkins workspace.)

Build artifacts
---------------

In the job/project:

* Post-build actions > Archive the artifacts
* Files to archive: runs/

stbt scripts
------------

Install the plugin for your source-control system (git, svn, etc.).

Configure the Jenkins job to checkout the repository containing your
stbt test scripts into a subdirectory of the workspace (so it won't
conflict with files generated by jenkins-stbt-run in the workspace).

SLAVE CONFIGURATION
===================

The slave must have a user (say, "stb-tester") that Jenkins connects as
(specified in the Jenkins "nodes" configuration). This user's
~/.ssh/authorized_keys must include Jenkins' public key so that the
Jenkins master can connect to the slave without entering a password.
Test this configuration by `ssh`ing from the Jenkins master to the
slave; you have to do this at least once to avoid the warning "the
authenticity of host ... can't be established. Are you sure you want to
continue connecting?".

The user that Jenkins connects as must have permission to read from
/dev/video0 -- on my Linux distribution I had to add the user to group
"video".

The slave must have `stbt` installed and available on the PATH.

The slave must have `jenkins-stbt-run` available. It doesn't have to
be on the PATH because when you configure the Jenkins job you can give
a full path to `jenkins-stbt-run`.

The slave must have a `stbt.conf` with the correct configuration for
stbt (the correct gstreamer source pipeline, the set-top box's IP
address if using the virtual remote control, etc.). Set `sink_pipeline =
fakesink`.

You can place this config file in ~/.config/stbt/stbt.conf or else in
any location you like and set STBT_CONFIG_FILE in the Jenkins job
configuration.

Note that I had to set DISPLAY=:0.0, and run "xhost +" on the slave,
even though stbt.conf configures "sink_pipeline = fakesink" (i.e. no
video is shown on the slave's screen) because stbt.py's "import gtk"
actually requires a connection to the X display. (This is something that
should be fixed in stb-tester.)

The slave must be able to connect to your source-control server (to
checkout or update the repository containing the test scripts) without
a password.