README: stbt no longer requires gstreamer-plugins-bad
In commit 0d588c3 we removed the dependency on gst-plugins-bad by
placing a copy of the templatematch element in the stb-tester repository
itself: We now build templatematch (renamed to "stbt-templatematch")
into the gstreamer plugin "libgst-stb-tester.so".

This commit merely updates the documentation to suit.

The version requirements for OpenCV have been relaxed; now that we don't
have to build *all* of gst-plugins-bad we can build successfully against
OpenCV 2.4.