'stbt run' times out when the gstreamer pipeline fails to start
Add a timeout to prevent stbt run from hanging when the gtk main loop is
started but nothing is received from the bus.

Note that the timeout is now effective from the call to the gtk main
loop, and not from the first buffer received (so the timeout may need to
be set slightly (1 second) longer?)

Note that when we restart the pipeline on underrun, we now reset the
wait_for_match / wait_for_motion timeout.