templatematch: Make the matching algorithm more robust
In some cases, the algorithm reported false positives or false negatives.

False positives were reported with large template images that differed in
small parts of the source images. The second pass of the algorithm was not
executed as the first pass of the algorithm provided a great confidence.
To solve this problem, the second pass is now always executed.

False negatives were reported with large template images (at least 100px on
one side) that had some parts with no edges detected. The template match
did not report a match when comparing two black images.
Other false positives were reported by the algorithm with large template
images that differed in small parts of the source images. This depended on
the split points chosen by the second pass.
To solve these two problems, the second pass has been improved and is now
much simpler. The absolute difference between the template image and the
source image is computed. To account for noise, the result is thresholded
and eroded. If the template is different enough, some white blobs will
remain.