templatematch: Cleanup - Make the code clearer
Note that when setting a new template image, the lock to filter is now held
a bit longer.

Holding the lock longer is not important as the only other part of the code
that could be delayed is the template match algorithm. It is fair to assume
that when setting a new template image, we are not concerned about delaying
a templatematch that would have otherwise used a non interesting template.