templatematch: template_match 'result' is now 'match' and is a boolean
It is now clearer that the result of the algorithm is match or no match.

It was misleading to have a floating point value returned for the result
of the match as there is now only two possible outcomes.
Either the first pass finds a match and the second pass confirms it, or no
match is found.

Now that the template match result is boolean, it makes more sense for the
visualisation to display a red rectangle in case of match and a yellow
rectangle otherwise.