wait_for_match and press_until_match: add optional parameter noise_threshold
The parameter noise_threshold enables to specify the threshold to use to
confirm a match by setting the property noiseThreshold in
stbt-templatematch.

When a match is detected with the first pass of the algorithm, the second
pass confirms the match by computing the absolute difference between the
source image and the template image. In order to account for noise, the
result is thresholded and it is now possible to specify the value of this
threshold. The default value is equivalent to a threshold of 41 and the
threshold was previously 40.