Rename C (GObject) class GstTemplateMatch to StbtTemplateMatch
to avoid conflicts with the templatematch element libgstopencv.so from
gst-plugins-bad (on some systems, gstreamer-plugins-bad-free-extras).

Even though the name of the element (as visible to users through
gst-inspect, gst-launch, etc) was "stbt-templatematch", in the C code
the name was still "GstTemplateMatch" which caused a conflict with the
templatematch element from libgstopencv.so.

Before this fix, the workaround was to remove the gstreamer registry
(~/.gstreamer-0.10/registry.*), uninstall libgstopencv.so, run
`gst-inspect stbt-templatematch`, and then re-install libgstopencv.so;
thereafter the templatematch element from libgstopencv.so was not
available but at least the other elements from that plugin were
available. Now, with the fix from this commit, no such workaround is
necessary, and both stbt-templatematch and templatematch are available.

Also renamed GstMotionDetect to StbtMotionDetect for consistency.