stbt: Refactor wait_for_match and wait_for_motion
This is to make it easier for a new feature that gives access to the list
of results that stbt-templatematch and stbt-motiondetect return. It also
cleans up the code and could make it possible to have the main loop running
in the background.

We would benefit from having a motion api that can return a list of
[timestamp, motion or not].
We would probably also benefit from having an api that can return a list of
[timestamp, match or not, position, name of template]. This would allow
to measure the smoothness of animation, wait for a match in a specific
position.

wait_for_match and wait_for_motion are simplified and the logic is clearer.

If necessary, with a minimum refactoring, it will also allow more easily to
have the main loop running in the background due to the use of the
synchronized queue. (see http://docs.python.org/library/queue.html)