stbt: Expose detect_motion and detect_match to the tests scripts
This makes it possible for the tests scripts to call the two iterator
methods detect_motion and detect_match.

This gives access to everything returned by the gstreamer elements
stbt-templatematch and stbt-motiondetect. These two methods are factored
out of Display and a single method detect is implemented. This cleans up
considerably the code and avoids having a class Display responsible for
doing everything.