`stbt run` times out using video stream time to fix the race condition
This fixes the race condition exposed by f289d24.

The timeout implementation is now based on the video stream time instead of
the real time. It will make it possible to implement process_all_frames by
slowing down the pipeline without having to change the timeout values in
the test scripts.

The gtk timeout added in 58285fb is kept but times out after 5 seconds. It
is now explicitly cancelled after the first message is received on the bus.
This is because this timeout is only used to prevent stbt run from hanging
when the gtk main loop is started but nothing is received on the bus.
Note that there are no tests that prove that this timeout is useful.

Note also that now, stbt explicitly fails after 3 consecutive underruns
whereas it used to rely on the gtk timeout to occurs after the third
underrun.